# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Allbtc Swagger UI",
        "description": "Документация OpenApi для Allbtc pool",
        "contact": {
            "email": "ognevoy.bima@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/v1/"
        }
    ],
    "paths": {
...
        "/referrals/incomes/{user}": {
            "get": {
                "tags": [
                    "Referral"
                ],
                "summary": "Get referral incomes list for a user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "daily_amount": {
                                                        "type": "string"
                                                    },
                                                    "hash": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "worker_count": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/referrals/{user}": {
            "get": {
                "tags": [
                    "Referral"
                ],
                "summary": "Get referral subs for a user",
                "operationId": "df834cce778d87e0953d60ee172b9889",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "referral_subs": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReferralResourceCollection"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/referrals/statistic/{user}": {
            "get": {
                "tags": [
                    "Referral"
                ],
                "summary": "Get referral statistics for a user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statistic": {
                                            "$ref": "#/components/schemas/ReferralStatisticResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/statistic/{sub}": {
            "get": {
                "tags": [
                    "Subaccount"
                ],
                "summary": "Get statistics for a sub",
                "parameters": [
                    {
                        "name": "sub",
                        "in": "path",
                        "description": "Sub's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for hash rate data",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "hashes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/HashRateResource"
                                            }
                                        },
                                        "incomes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/IncomeCollection"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Sub not found"
                    }
                }
            }
        },
...
        "/wallets/{sub}": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Get wallets for a sub",
                "parameters": [
                    {
                        "name": "sub",
                        "in": "path",
                        "description": "Sub's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WalletResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Sub not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HashRateResource": {
                "properties": {
                    "hash": {
                        "type": "integer"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "worker_count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "IncomeCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "referral_id": {
                                    "type": "integer"
                                },
                                "amount": {
                                    "type": "number"
                                },
                                "hash": {
                                    "type": "integer"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "created_at": {
                                    "type": "string"
                                },
                                "updated_at": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ReferralResourceCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "referral_active_workers_count": {
                                    "type": "integer"
                                },
                                "referral_inactive_workers_count": {
                                    "type": "integer"
                                },
                                "referral_hash_per_day": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "total_amount": {
                                    "type": "number",
                                    "format": "float"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ReferralStatisticResource": {
                "properties": {
                    "group_id": {
                        "type": "integer"
                    },
                    "attached_referrals_count": {
                        "type": "integer"
                    },
                    "active_referrals_count": {
                        "type": "integer"
                    },
                    "referrals_total_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "code": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "WalletResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "percent": {
                        "type": "number",
                        "format": "float"
                    },
                    "minWithdrawal": {
                        "type": "number",
                        "format": "float"
                    },
                    "wallet": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "total_payout": {
                        "type": "number",
                        "format": "float"
                    },
                    "is_unlocked": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    }
}